Move the settings handlers and the updates socket out of App.vue
The sidebar and settings-dialog handlers become useAppSettingsHandlers -
thin validate-and-write functions, with the persisting left to the config
watchers - and the /updates WebSocket, its filter handshake, the reconnect
loop and the realtime-sync wiring become useUpdatesSocket. App.vue still
owns the socket's lifecycle: it connects on mount and disconnects on
unmount, through the composable.

App.vue drops from 1405 to 1104 lines.
